requirements.txt
flask==2.3.2
flask-smorest==0.42.0
flask-socketio==5.3.4
flask-cors==3.0.10
flask-sqlalchemy==3.0.5
prometheus-client==0.16.0
requests==2.31.0
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
